Second Update to configuration
Updated a large chunk of the configuration to have "is any" matches for exact file names and "contains any" for folder paths and heuristic matches.

This is being done  to reduce the number of partial matches from 2-3 character binaries. Once im done with this it should reduce false positive logging events overall.