Fixed logic errors with "is any" filters
Spoke with @Cyb3rMonk and identified  a misunderstanding on how the "is any" filter works in Sysmon. Replaced it with contains any but prepended a \ to the front of the imagenames to filter in a similar intended manner. Turns out that "is any" requires matching the file path it isn't full path OR filename.

Also, mirrored the imageloaded section to the  driverload section. This should hopefully catch driver load events from known process names. Later I will revisit this section and identify which drivers each installation of the RMMs install to include manually in this section.